Coord: Moved restart script code into restartscript.cpp.
